[3pt] - Create feature: add new like - Student A#14
